feat: country form component
Also moved a single line in the member form component.
